Initial implementation of the proxy server.
Depending on where the configuration data will live, this might already be
sufficient.

I chose http-proxy-middleware, because it uses http-proxy under the hood, which
I've used extensively in the past and regard as the most robust NodeJS proxy.
I chose the `/cid/:cid` endpoint/ middleware hook, because we also need to
support the obligatory Mozilla Service routes, like `__heartbeat__`.
The configuration is supposed to be sent over to the Docker container as
environment variable and is expected to be a valid JSON string, keyed by
campaign ID.