Bug 1680019 - Add weather endpoint
We use two different weather endpoints in the extension, so we need two different endpoints on the server. An early version of this patch used a single endpoint (/cid/weather) and passed the pathname of the specific endpoint in an X-Path header. The problem with this is that the client only saw repeated requests to `https://topsites.mozilla.com/cid/weather`, so it would cache data far too aggressively. We need to actually call different paths using different query parameters to get appropriate caching behaviour.

As alluded to above, this change also adds support for clients to pass query paramters in their requests to the server. These query paramters are then merged with the params pre-defined in CONFIG. For example, consider a request to `https://topsites.mozilla.com/cid/weather_location?q=“toronto`. The `q` parameter is merged with the `apikey` parameter predefined in CONFIG to create a final URL of
`https://<WEATHER_LOCATION_URL>?q=toronto&apikey=abc`